エネミーのHPの表示できるようにした

+++ Assets/Script/Enemys/Boss/BossMove.cs
@@ -146,9 +146,9 @@ public class BossMove : MonoBehaviour
     }
 
 
-    public float HpGet()
+    public int HpGet()
     {
-        return enemyHP;
+        return (int)enemyHP;
     }
